requirements.txt
# Canadian Market Opportunity Detection Agent
# requirements.txt
# Tested on Python 3.10+ / Ubuntu 20.04 ARM64 (Jetson Nano)

# HTTP and RSS
requests>=2.31.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0
lxml>=4.9.0

# YAML config
PyYAML>=6.0.1

# System resource monitoring
psutil>=5.9.8

# Data utilities (lightweight)
python-dateutil>=2.8.2

# Optional: FinBERT-style sentiment (uncomment if GPU/model available)
# transformers>=4.35.0
# torch>=2.0.0

# Development / testing
pytest>=7.4.0
pytest-cov>=4.1.0
